moved usecases into either the access or owner sub packages
